reverse Grid argument order again
Apparently the order of arguments for standard library containers
is not as consistent as I first thought. Most immutable container
functions take the container type as last argument.
Some mutable containers (Stack, Queue) do this as well, but Hashtbl
takes the main argument first.
For Grid I initially also placed the main argument first, because
it "felt best", but changed it to align with List and Array.
But Grid is mutable, so it would be more fitting to align it with
Hashtbl instead.